race condition between mkdir and stat
Under heavy load whenever there are many parallel calls to `mkdirp` and
directory is constantly being created and deleted, it's possible for
`mkdir` to fail with `EEXIST` and for `stat` to fail with `ENOENT`
instead of actual value. It's better to re-try `mkdir`/`stat` as a
whole, not just `stat` alone, as parent directory might be gone and we
would end up with an infinite loop.